Move `OntologyOutwardEdge` to `utoipa_typedef`
