+++ apps/hash-graph/lib/graph/src/api/rest/utoipa_typedef/subgraph.rs
@@ -5,7 +5,10 @@ use serde::Serialize;
 use utoipa::ToSchema;
 
 pub use self::{
-    edges::{Edges, KnowledgeGraphOutwardEdge, KnowledgeGraphRootedEdges, OntologyRootedEdges},
+    edges::{
+        Edges, KnowledgeGraphOutwardEdge, KnowledgeGraphRootedEdges, OntologyOutwardEdge,
+        OntologyRootedEdges,
+    },
     vertices::{
         KnowledgeGraphVertex, KnowledgeGraphVertices, OntologyVertex, OntologyVertices, Vertex,
         Vertices,
