+++ apps/hash-graph/lib/graph/src/api/rest/utoipa_typedef/subgraph/edges.rs
@@ -10,11 +10,53 @@ use utoipa::{
 use crate::{
     identifier::{knowledge::EntityId, ontology::OntologyTypeVersion, time::Timestamp},
     subgraph::{
-        edges::{KnowledgeGraphEdgeKind, OntologyOutwardEdge, OutwardEdge, SharedEdgeKind},
+        edges::{KnowledgeGraphEdgeKind, OntologyEdgeKind, OutwardEdge, SharedEdgeKind},
         identifier::{EntityIdWithInterval, OntologyTypeVertexId},
         temporal_axes::VariableAxis,
     },
 };
+
+#[derive(Debug, Hash, PartialEq, Eq, Serialize)]
+#[serde(untagged)]
+pub enum OntologyOutwardEdge {
+    ToOntology(OutwardEdge<OntologyEdgeKind, OntologyTypeVertexId>),
+    ToKnowledgeGraph(OutwardEdge<SharedEdgeKind, EntityIdWithInterval>),
+}
+
+impl From<OutwardEdge<OntologyEdgeKind, OntologyTypeVertexId>> for OntologyOutwardEdge {
+    fn from(edge: OutwardEdge<OntologyEdgeKind, OntologyTypeVertexId>) -> Self {
+        Self::ToOntology(edge)
+    }
+}
+
+impl From<OutwardEdge<SharedEdgeKind, EntityIdWithInterval>> for OntologyOutwardEdge {
+    fn from(edge: OutwardEdge<SharedEdgeKind, EntityIdWithInterval>) -> Self {
+        Self::ToKnowledgeGraph(edge)
+    }
+}
+
+// WARNING: This MUST be kept up to date with the enum variants.
+//   We have to do this because utoipa doesn't understand serde untagged:
+//   https://github.com/juhaku/utoipa/issues/320
+impl ToSchema<'_> for OntologyOutwardEdge {
+    fn schema() -> (&'static str, RefOr<Schema>) {
+        (
+            "OntologyOutwardEdge",
+            OneOfBuilder::new()
+                .item(
+                    <OutwardEdge<OntologyEdgeKind, OntologyTypeVertexId>>::generate_schema(
+                        "OntologyToOntologyOutwardEdge",
+                    ),
+                )
+                .item(
+                    <OutwardEdge<SharedEdgeKind, EntityIdWithInterval>>::generate_schema(
+                        "OntologyToKnowledgeGraphOutwardEdge",
+                    ),
+                )
+                .into(),
+        )
+    }
+}
 
 #[derive(Debug, Serialize)]
 #[serde(untagged)]
