+++ apps/hash-graph/lib/graph/src/subgraph/edges/edge.rs
@@ -1,11 +1,6 @@
 use serde::Serialize;
 use utoipa::{openapi, ToSchema};
 
-use crate::subgraph::{
-    edges::{KnowledgeGraphEdgeKind, OntologyEdgeKind, SharedEdgeKind},
-    identifier::{EntityIdWithInterval, OntologyTypeVertexId},
-};
-
 #[derive(Debug, Hash, PartialEq, Eq, Serialize)]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
 pub struct OutwardEdge<K, E> {
@@ -40,53 +35,4 @@ where
             .required("rightEndpoint")
             .into()
     }
-}
-
-#[derive(Debug, Hash, PartialEq, Eq, Serialize)]
-#[serde(untagged)]
-pub enum OntologyOutwardEdge {
-    ToOntology(OutwardEdge<OntologyEdgeKind, OntologyTypeVertexId>),
-    ToKnowledgeGraph(OutwardEdge<SharedEdgeKind, EntityIdWithInterval>),
-}
-
-impl From<OutwardEdge<OntologyEdgeKind, OntologyTypeVertexId>> for OntologyOutwardEdge {
-    fn from(edge: OutwardEdge<OntologyEdgeKind, OntologyTypeVertexId>) -> Self {
-        Self::ToOntology(edge)
-    }
-}
-
-impl From<OutwardEdge<SharedEdgeKind, EntityIdWithInterval>> for OntologyOutwardEdge {
-    fn from(edge: OutwardEdge<SharedEdgeKind, EntityIdWithInterval>) -> Self {
-        Self::ToKnowledgeGraph(edge)
-    }
-}
-
-// WARNING: This MUST be kept up to date with the enum variants.
-//   We have to do this because utoipa doesn't understand serde untagged:
-//   https://github.com/juhaku/utoipa/issues/320
-impl ToSchema<'_> for OntologyOutwardEdge {
-    fn schema() -> (&'static str, openapi::RefOr<openapi::Schema>) {
-        (
-            "OntologyOutwardEdge",
-            openapi::OneOfBuilder::new()
-                .item(
-                    <OutwardEdge<OntologyEdgeKind, OntologyTypeVertexId>>::generate_schema(
-                        "OntologyToOntologyOutwardEdge",
-                    ),
-                )
-                .item(
-                    <OutwardEdge<SharedEdgeKind, EntityIdWithInterval>>::generate_schema(
-                        "OntologyToKnowledgeGraphOutwardEdge",
-                    ),
-                )
-                .into(),
-        )
-    }
-}
-
-#[derive(Debug, Hash, PartialEq, Eq, Serialize)]
-#[serde(untagged)]
-pub enum KnowledgeGraphOutwardEdge {
-    ToKnowledgeGraph(OutwardEdge<KnowledgeGraphEdgeKind, EntityIdWithInterval>),
-    ToOntology(OutwardEdge<SharedEdgeKind, OntologyTypeVertexId>),
 }
